+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change log
 ==========
 
+## Version 1.4.0 (2014-04-07)
+
+Features:
+
+  - [#8](https://github.com/geoffroy-aubry/Supervisor/issues/8): Allow to force supervisor execution_id.
+  - [#7](https://github.com/geoffroy-aubry/Supervisor/issues/7): Add proper signal handling and interrupts.
+
 ## Version 1.3.1 (2013-11-11)
 
 Features: